src/requirements.txt
#functions-framework==3.*
google-cloud-compute
SQLAlchemy==2.0.21
sqlalchemy-pytds==1.0.0     
PyMySQL==1.1.0
pg8000==1.30.2
asyncpg==0.28.0
python-tds==1.13.0
cloud-sql-python-connector